Implemented Data Passing
Going from client details to project create passes through the client info.
Same with project details to bid create.

Minor bug: When first loading in the client/project info in the drop down list; it does not show the other options when clicking on the drop down until you update the search text.